+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Optical flow method as a refinement possible step
+- Phase cross correlation as an estimation possible step
 
 ## 0.1.0 (february 2022)